+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1.0
 
+* Add `MemPack` instance for `Void`
 * Add `packWithByteArray` and `packWithMutableByteArray`
 * Fix infinite loop during list decoding when length encoded was negative